// src/app/models/restaurant.model.ts
export interface Restaurant {
    resId: number;
    resName: string;
    registrationId: string;
    resBranch: string;
    resType: string;
    address: string;
    canton: string;
    district: string;
    province: string;
    zipcode: number;
    resTelephone: string;
    resEmail: string;
    resWebsite: string;
    resImage: string;
    description: string;
    resOpen: String;
    resClose: String;
    imageURL: String;
    
    retrievedImage: string;
    isAdded: boolean;
    
}